refactor(shop): ProductAdminService에서 파일 처리 로직 분리
- 파일 업로드 로직을 FileUploadService로 분리하여 단일 책임 원칙을 적용
- 파일 처리 실패 시 커스텀 예외(ImageUploadException)를 던지도록 수정하여 데이터 일관성을 보장
